Comments from Kathy
Hope this is a workable way to give comments, please advise if not: My comments are inserted into the file as <!-- #KC_YYYY-MM-DD: comment --> so you can search for them.  I've consulted with Carolyn so we are not duplicating things.  Diane, sorry I didn't notice your header had my initials backwards until now and the header is now in several files. I sent an email response in November about that (during the initial request for feedback on the header) and assumed it was revised.